+++ go.work
@@ -1,9 +1,12 @@
-go 1.20
+go 1.21
 
 use (
 	.
 	./cmd
 	./launcher
+	./verifier
 )
 
-replace github.com/google/go-tpm-tools v0.3.12 => ./
+replace github.com/google/go-tpm-tools v0.4.4 => ./
+
+replace github.com/google/go-tpm-tools/verifier v0.4.4 => ./
