+++ launcher/launcher/main.go
@@ -53,6 +53,7 @@ var exitMessage = "TEE container launcher exiting"
 func main() {
 	var exitCode int // by default exit code is 0
 	var err error
+	ctx := context.Background()
 
 	logger = log.Default()
 	// log.Default() outputs to stderr; change to stdout.
@@ -82,7 +83,7 @@ func main() {
 
 	// Get RestartPolicy and IsHardened from spec
 	mdsClient = metadata.NewClient(nil)
-	launchSpec, err := spec.GetLaunchSpec(mdsClient)
+	launchSpec, err := spec.GetLaunchSpec(ctx, mdsClient)
 	if err != nil {
 		logger.Printf("failed to get launchspec, make sure you're running inside a GCE VM: %v\n", err)
 		// if cannot get launchSpec, exit directly
@@ -122,7 +123,7 @@ func main() {
 			logger.Printf("%s, exit code: %d\n", exitMessage, exitCode)
 		}
 	}()
-	if err = startLauncher(launchSpec, serialConsole); err != nil {
+	if err = startLauncher(ctx, launchSpec, serialConsole); err != nil {
 		logger.Println(err)
 	}
 
@@ -161,7 +162,7 @@ func getExitCode(isHardened bool, restartPolicy spec.RestartPolicy, err error) i
 	return exitCode
 }
 
-func startLauncher(launchSpec spec.LaunchSpec, serialConsole *os.File) error {
+func startLauncher(ctx context.Context, launchSpec spec.LaunchSpec, serialConsole *os.File) error {
 	logger.Printf("Launch Spec: %+v\n", launchSpec)
 	containerdClient, err := containerd.New(defaults.DefaultAddress)
 	if err != nil {
@@ -185,12 +186,12 @@ func startLauncher(launchSpec spec.LaunchSpec, serialConsole *os.File) error {
 	}
 	gceAk.Close()
 
-	token, err := launcher.RetrieveAuthToken(mdsClient)
+	token, err := launcher.RetrieveAuthToken(ctx, mdsClient)
 	if err != nil {
 		logger.Printf("failed to retrieve auth token: %v, using empty auth for image pulling\n", err)
 	}
 
-	ctx := namespaces.WithNamespace(context.Background(), namespaces.Default)
+	ctx = namespaces.WithNamespace(ctx, namespaces.Default)
 	r, err := launcher.NewRunner(ctx, containerdClient, token, launchSpec, mdsClient, tpm, logger, serialConsole)
 	if err != nil {
 		return err
